tests/wpt/web-platform-tests/css/css-transitions/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of CSS Transitions properties</title>
<link rel="help" href="https://drafts.csswg.org/css-transitions/#property-index">
<meta name="assert" content="Properties inherit or not according to the spec.">
<meta name="assert" content="Properties have initial values according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
<div id="target"></div>
</div>
<script>
assert_not_inherited('transition-delay', '0s', '2s');
assert_not_inherited('transition-duration', '0s', '3s');
assert_not_inherited('transition-property', 'all', 'opacity');
assert_not_inherited('transition-timing-function', 'ease', 'linear');
</script>
</body>
</html>
